+++ pasqal_cloud/errors.py
@@ -8,7 +8,12 @@ class ExceptionWithResponseContext(BaseException):
     def __init__(self, msg: str, e: Optional[HTTPError] = None) -> None:
         if not e:
             return super().__init__(msg)
+
+        if e.response is None:
+            return super().__init__(msg)
+
         resp: Response = e.response
+
         if not resp.content:
             return super().__init__(msg)
         data = resp.json()