+++ tests/test_device_specs.py
@@ -8,6 +8,8 @@ from pasqal_cloud import SDK
 from pasqal_cloud.errors import DeviceSpecsFetchingError
 from tests.test_doubles.authentication import FakeAuth0AuthenticationSuccess
 
+from typing import Any, Generator
+
 
 class TestDeviceSpecs:
     @pytest.fixture(autouse=True)
@@ -17,6 +19,15 @@ class TestDeviceSpecs:
             username="me@test.com", password="password", project_id=str(uuid4())
         )
 
+    @pytest.fixture(autouse=True)
+    def mock_sleep(self):
+        """
+        This fixture overrides sleeps, so tests don't need to wait for
+        the total elapsed time.
+        """
+        with patch("time.sleep"):
+            yield
+
     @pytest.mark.usefixtures("mock_request")
     def test_get_device_specs_success(self):
         device_specs_dict = self.sdk.get_device_specs_dict()
@@ -24,7 +35,9 @@ class TestDeviceSpecs:
         specs = device_specs_dict["FRESNEL"]
         json.loads(specs)
 
-    def test_get_device_specs_error(self, mock_request_exception):
+    def test_get_device_specs_error(
+        self, mock_request_exception: Generator[Any, Any, None]
+    ):
         with pytest.raises(DeviceSpecsFetchingError):
             _ = self.sdk.get_device_specs_dict()
         assert mock_request_exception.last_request.method == "GET"